同学圈详情界面

+++ app/src/main/java/com/tanlifei/app/common/utils/NumberUtils.kt
@@ -11,19 +11,19 @@ class NumberUtils {
     companion object {
         private const val MAX_SHOW_PRAISE_COUNT = 10000.0
 
-        fun setPraiseCount(count: Int): String? {
+        fun setPraiseCount(count: Long): String? {
             return showCount("点赞", count)
         }
 
-        fun setCommentCount(defaultStr: String, count: Int): String? {
+        fun setCommentCount(defaultStr: String, count: Long): String? {
             return showCount(defaultStr, count)
         }
 
-        fun setCommentCount(count: Int): String? {
+        fun setCommentCount(count: Long): String? {
             return showCount("评论", count)
         }
 
-        private fun showCount(defaultStr: String, count: Int): String {
+        private fun showCount(defaultStr: String, count: Long): String {
             return if (count <= 0) {
                 defaultStr
             } else if (count > 0 && count < MAX_SHOW_PRAISE_COUNT) {
